+++ extensions/unix-timestamp/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## QoL improvements (2024-07-14)
+- Allow input timestamp from arguments
+
 ## QoL improvements (2022-03-15)
 - Show HUD instead of showing the toast
 - Better input validation